pass: support extensions with `$out/bin`
An example for that would be `passExtensions.pass-import` where
`pimport` is available in `$out/bin`. In that case, `$out/bin` in
`buildEnv` isn't a symlink anymore and doesn't need to be removed.

+++ pkgs/tools/security/pass/default.nix
@@ -34,11 +34,15 @@ let
 
       postBuild = ''
         files=$(find $out/bin/ -type f -exec readlink -f {} \;)
-        rm $out/bin
-        mkdir $out/bin
+        if [ -L $out/bin ]; then
+          rm $out/bin
+          mkdir $out/bin
+        fi
 
         for i in $files; do
-          ln -sf $i $out/bin/$(basename $i)
+          if ! [ "$(readlink -f "$out/bin/$(basename $i)")" = "$i" ]; then
+            ln -sf $i $out/bin/$(basename $i)
+          fi
         done
 
         wrapProgram $out/bin/pass \
